Rename fixtures/config_files to fixtures/config
